refactor(edge): complete V3 Lite closeout
