--- .github/workflows/pr94-clean-final-20260819.yml
@@ -1,90 +1,0 @@
-name: PR94 clean final state
-
-on:
-  push:
-    branches:
-      - agent/v3-lite-edge-presentation
-    paths:
-      - scripts/pr94-clean-final/TRIGGER
-
-permissions:
-  contents: write
-  pull-requests: write
-  issues: write
-
-jobs:
-  verify-and-clean:
-    runs-on: windows-latest
-    timeout-minutes: 30
-    defaults:
-      run:
-        shell: pwsh
-    env:
-      DOTNET_CLI_TELEMETRY_OPTOUT: "1"
-      DOTNET_NOLOGO: "true"
-    steps:
-      - name: Check out current PR94
-        uses: actions/checkout@v7
-        with:
-          ref: agent/v3-lite-edge-presentation
-          fetch-depth: 0
-          submodules: recursive
-
-      - name: Set up .NET
-        uses: actions/setup-dotnet@v6
-        with:
-          dotnet-version: "10.0.x"
-
-      - name: Verify final architecture
-        run: |
-          if (-not (Test-Path EdgeCapsuleDockingCompositionBridge.cs)) { throw "EdgeCapsuleDockingCompositionBridge.cs missing" }
-          if (-not (Select-String -Path EdgeCapsuleDockingCompositionBridge.cs -Pattern "CreateSurfaceFromHwnd" -Quiet)) { throw "live-surface bridge marker missing" }
-          if (Select-String -Path EdgeCapsuleDockingCompositionBridge.cs -Pattern "SetClip|SetScale|EffectGroup|RenderTargetBitmap|Snapshot" -Quiet) { throw "forbidden docking capability present" }
-          if (-not (Select-String -Path EdgeCapsuleDragWindow.cs -Pattern "EdgeCapsuleDockingCompositionBridge" -Quiet)) { throw "drag window does not route through docking bridge" }
-          git -c core.whitespace=cr-at-eol diff --check
-
-      - name: Restore
-        run: |
-          dotnet restore .\PaperTodo.csproj
-          dotnet restore .\scripts\edge-refinement-tests\EdgeRefinementChecks.csproj
-
-      - name: Debug build
-        run: dotnet build .\PaperTodo.csproj -c Debug --no-restore
-
-      - name: Release build
-        run: dotnet build .\PaperTodo.csproj -c Release --no-restore
-
-      - name: Edge refinement checks
-        run: dotnet run --project .\scripts\edge-refinement-tests\EdgeRefinementChecks.csproj -c Release --no-restore
-
-      - name: Remove temporary PR94 infrastructure
-        run: |
-          git rm -r --ignore-unmatch .github/workflows/pr94-source-export.yml scripts/pr94-source-export
-          git rm -r --ignore-unmatch .github/workflows/pr94-successor-hardening.yml scripts/pr94-hardening
-          git rm -r --ignore-unmatch .github/workflows/pr94-finalize.yml scripts/pr94-finalize
-          git rm --ignore-unmatch .github/workflows/pr94-finalize-20260819.yml
-          git rm -r --ignore-unmatch .github/workflows/pr94-clean-final-20260819.yml scripts/pr94-clean-final
-
-      - name: Publish clean final tree
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add -A
-          git -c core.whitespace=cr-at-eol diff --cached --check
-          if (-not (git diff --cached --quiet)) {
-            git commit -m "chore(edge): finalize PR94 clean state"
-            git push origin HEAD:agent/v3-lite-edge-presentation
-          }
-
-      - name: Report clean success
-        env:
-          GH_TOKEN: ${{ github.token }}
-        run: |
-          $head = git rev-parse HEAD
-          gh pr comment 94 --repo snownico0722/PaperTodo --body "### PR94 最终净化验收通过`n`n- head: ``$head```n- Debug Build: success`n- Release Build: success`n- Edge checks: success`n- DComp docking bridge: present`n- 临时 export/hardening/finalizer 文件: removed`n- run: $env:GITHUB_SERVER_URL/$env:GITHUB_REPOSITORY/actions/runs/$env:GITHUB_RUN_ID"
-
-      - name: Report failure
-        if: failure()
-        env:
-          GH_TOKEN: ${{ github.token }}
-        run: gh pr comment 94 --repo snownico0722/PaperTodo --body "### PR94 最终净化验收失败，未发布清理提交`n`nrun: $env:GITHUB_SERVER_URL/$env:GITHUB_REPOSITORY/actions/runs/$env:GITHUB_RUN_ID"
