--- .github/workflows/pr94-finalize-20260819.yml
@@ -1,114 +1,0 @@
-name: PR94 final V3 Lite one-shot
-
-on:
-  push:
-    branches:
-      - agent/v3-lite-edge-presentation
-    paths:
-      - scripts/pr94-finalize/TRIGGER_20260819
-
-permissions:
-  contents: write
-  pull-requests: write
-  issues: write
-
-jobs:
-  finalize:
-    runs-on: windows-latest
-    timeout-minutes: 35
-    defaults:
-      run:
-        shell: pwsh
-    env:
-      DOTNET_CLI_TELEMETRY_OPTOUT: "1"
-      DOTNET_NOLOGO: "true"
-    steps:
-      - name: Check out PR94
-        uses: actions/checkout@v7
-        with:
-          ref: agent/v3-lite-edge-presentation
-          fetch-depth: 0
-          submodules: recursive
-
-      - name: Set up .NET
-        uses: actions/setup-dotnet@v6
-        with:
-          dotnet-version: "10.0.x"
-
-      - name: Apply reviewed final implementation when needed
-        run: |
-          if (Test-Path EdgeCapsuleDockingCompositionBridge.cs) {
-            Write-Host "Final docking bridge already present; validating current source."
-            exit 0
-          }
-          if (-not (Test-Path scripts/pr94-finalize/final.patch.xz.b64)) {
-            throw "Missing reviewed final.patch.xz.b64"
-          }
-          @'
-          import base64
-          import lzma
-          from pathlib import Path
-
-          encoded = Path("scripts/pr94-finalize/final.patch.xz.b64").read_text(encoding="utf-8").strip()
-          patch = lzma.decompress(base64.b64decode(encoded))
-          Path("scripts/pr94-finalize/final.patch").write_bytes(patch)
-          '@ | python -
-          git apply --whitespace=nowarn scripts/pr94-finalize/final.patch
-
-      - name: Verify architecture shape
-        run: |
-          git -c core.whitespace=cr-at-eol diff --check
-          if (-not (Test-Path EdgeCapsuleDockingCompositionBridge.cs)) { throw "docking bridge missing" }
-          if (-not (Select-String -Path EdgeCapsuleDockingCompositionBridge.cs -Pattern "CreateSurfaceFromHwnd" -Quiet)) { throw "live HWND surface marker missing" }
-          if (Select-String -Path EdgeCapsuleDockingCompositionBridge.cs -Pattern "SetClip|SetScale|EffectGroup|RenderTargetBitmap|Snapshot" -Quiet) { throw "forbidden compositor capability returned" }
-          if (Select-String -Path EdgeCapsuleDragWindow.cs -Pattern "TryMoveWindowDevicePosition\(" -Quiet) {
-            Write-Host "Native fallback remains available; normal-path regression checks must prove DComp-first routing."
-          }
-          git diff --stat
-
-      - name: Restore
-        run: |
-          dotnet restore .\PaperTodo.csproj
-          dotnet restore .\scripts\edge-refinement-tests\EdgeRefinementChecks.csproj
-
-      - name: Debug build
-        run: dotnet build .\PaperTodo.csproj -c Debug --no-restore
-
-      - name: Release build
-        run: dotnet build .\PaperTodo.csproj -c Release --no-restore
-
-      - name: Edge refinement checks
-        run: dotnet run --project .\scripts\edge-refinement-tests\EdgeRefinementChecks.csproj -c Release --no-restore
-
-      - name: Remove all temporary PR94 infrastructure
-        run: |
-          git rm -r --ignore-unmatch .github/workflows/pr94-source-export.yml scripts/pr94-source-export
-          git rm -r --ignore-unmatch .github/workflows/pr94-successor-hardening.yml scripts/pr94-hardening
-          git rm -r --ignore-unmatch .github/workflows/pr94-finalize.yml scripts/pr94-finalize
-          git rm --ignore-unmatch .github/workflows/pr94-finalize-20260819.yml
-
-      - name: Commit validated final implementation
-        run: |
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add -A
-          git -c core.whitespace=cr-at-eol diff --cached --check
-          if (git diff --cached --quiet) {
-            Write-Host "No source changes remain to publish."
-            exit 0
-          }
-          git commit -m "feat(edge): complete V3 Lite presentation"
-          git push origin HEAD:agent/v3-lite-edge-presentation
-
-      - name: Report success
-        env:
-          GH_TOKEN: ${{ github.token }}
-        run: |
-          $head = git rev-parse HEAD
-          gh pr comment 94 --repo snownico0722/PaperTodo --body "### PR94 V3 Lite 最终实现已验证并发布`n`n- head: ``$head```n- Debug Build: success`n- Release Build: success`n- Edge refinement checks: success`n- Docking: endpoint-once + DComp live-surface translation`n- 临时 finalizer/source-export/hardening 文件已清理`n- run: $env:GITHUB_SERVER_URL/$env:GITHUB_REPOSITORY/actions/runs/$env:GITHUB_RUN_ID"
-
-      - name: Report failure
-        if: failure()
-        env:
-          GH_TOKEN: ${{ github.token }}
-        run: gh pr comment 94 --repo snownico0722/PaperTodo --body "### PR94 V3 Lite 最终流水线失败，未发布半成品`n`nrun: $env:GITHUB_SERVER_URL/$env:GITHUB_REPOSITORY/actions/runs/$env:GITHUB_RUN_ID"
