--- .github/workflows/pr94-source-export.yml
@@ -1,47 +1,0 @@
-name: PR94 source export
-
-on:
-  push:
-    branches:
-      - agent/v3-lite-edge-presentation
-    paths:
-      - scripts/pr94-source-export/READY
-
-permissions:
-  contents: write
-
-jobs:
-  export:
-    runs-on: windows-latest
-    timeout-minutes: 10
-    defaults:
-      run:
-        shell: pwsh
-    steps:
-      - name: Check out PR94 source
-        uses: actions/checkout@v7
-        with:
-          ref: agent/v3-lite-edge-presentation
-          fetch-depth: 0
-          submodules: recursive
-
-      - name: Pack complete working tree
-        run: |
-          tar -a -c -f pr94-source.zip --exclude=.git --exclude=pr94-source.zip .
-
-      - name: Upload source snapshot
-        uses: actions/upload-artifact@v4
-        with:
-          name: pr94-source
-          path: pr94-source.zip
-          if-no-files-found: error
-          retention-days: 1
-
-      - name: Publish export run id
-        run: |
-          Set-Content -LiteralPath scripts/pr94-source-export/RUN_ID -Value $env:GITHUB_RUN_ID -NoNewline
-          git config user.name "github-actions[bot]"
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git add scripts/pr94-source-export/RUN_ID
-          git commit -m "chore: record PR94 source export run"
-          git push origin HEAD:agent/v3-lite-edge-presentation
