--- scripts/pr94-clean-final/TRIGGER
@@ -1,1 +1,0 @@
-verify final V3 Lite source and remove temporary infrastructure
