--- scripts/pr94-finalize/READY
@@ -1,1 +1,0 @@
-Direct ref trigger number two for final V3 Lite validation.
