--- scripts/pr94-finalize/TRIGGER_20260819
@@ -1,1 +1,0 @@
-apply and validate final V3 Lite docking bridge
