--- scripts/pr94-hardening/SUCCESSOR_READY
@@ -1,1 +1,0 @@
-apply successor hardening from d4af6af
