--- scripts/pr94-source-export/READY
@@ -1,2 +1,0 @@
-export current PR94 cold-start diagnostics source
-2026-08-19T15:50+08:00
